Rename "SickbedsSummaryCard" to "SickbedsUsedRateCard"
