+++ components/cards/SickbedsUsedRateCard.vue
